+++ src/api/layers/listener.layer.ts
@@ -45,6 +45,7 @@ declare global {
     onStreamChange: any;
     onIncomingCall: any;
     onAck: any;
+    onMessageEdit: any;
   }
 }
 
@@ -140,6 +141,21 @@ export class ListenerLayer extends ProfileLayer {
           if (!window['onAck'].exposed) {
             window.WAPI.waitNewAcknowledgements(window['onAck']);
             window['onAck'].exposed = true;
+          }
+        } catch (error) {
+          console.error(error);
+        }
+        try {
+          if (!window['onMessageEdit'].exposed) {
+            WPP.on('chat.msg_edited', (data) => {
+              const eventData = {
+                chat: data.chat,
+                id: data.id,
+                msg: WAPI.processMessageObj(data.msg, true, false),
+              };
+              window['onMessageEdit'](eventData);
+            });
+            window['onMessageEdit'].exposed = true;
           }
         } catch (error) {
           console.error(error);
@@ -397,7 +413,7 @@ export class ListenerLayer extends ProfileLayer {
   }
 
   /**
-   * @event Listens to messages acknowledgement Changes
+   * @event Listens to message acknowledgement changes
    * @returns Disposable object to stop the listening
    */
   public onAck(callback: (ack: Ack) => void) {
@@ -405,20 +421,30 @@ export class ListenerLayer extends ProfileLayer {
   }
 
   /**
-   * Escuta os eventos de Localização em tempo real de todos os chats
-   * @event Eventos de Localização em tempo real
-   * @param callback Função para ser executada quando houver alterações
-   * @returns Objeto descartável para parar de ouvir
+   * @event Listens to message edited changes
+   * @returns Disposable object to stop the listening
+   */
+  public onMessageEdit(
+    callback: (chat: Wid, id: string, msg: Message) => void
+  ) {
+    return this.registerEvent(ExposedFn.onMessageEdit, callback);
+  }
+
+  /**
+   * Listens to real-time location events of all chats
+   * @event Real-time location events
+   * @param callback Function to be executed when changes are made
+   * @returns Disposable object to stop listening
    */
   public onLiveLocation(callback: (liveLocationEvent: LiveLocation) => void): {
     dispose: () => void;
   };
   /**
-   * Escuta os eventos de Localização em tempo real
-   * @event Eventos de Localização em tempo real
-   * @param id Único ID ou lista de IDs de contatos para acompanhar a localização
-   * @param callback Função para ser executada quando houver alterações
-   * @returns Objeto descartável para parar de ouvir
+   * Listens to location events in real time
+   * @event Location events in real time
+   * @param id Unique ID or list of contact IDs to track location
+   * @param callback Function to be executed when changes are made
+   * @returns Disposable object to stop listening
    */
   public onLiveLocation(
     id: string | string[],
@@ -451,17 +477,17 @@ export class ListenerLayer extends ProfileLayer {
   }
 
   /**
-   * @event Listens to participants changed
-   * @param to callback
+   * @event Listens to participant changes
+   * @param callback Function to be executed when participant changes occur
    * @returns Stream of ParticipantEvent
    */
   public onParticipantsChanged(callback: (evData: ParticipantEvent) => void): {
     dispose: () => void;
   };
   /**
-   * @event Listens to participants changed
-   * @param to group id: xxxxx-yyyy@us.c
-   * @param to callback
+   * @event Listens to participant changes in a certain group
+   * @param groupId xxxxx-yyyy@us.c
+   * @param callback Function to be executed when participant changes occur
    * @returns Stream of ParticipantEvent
    */
   public onParticipantsChanged(
@@ -520,8 +546,8 @@ export class ListenerLayer extends ProfileLayer {
     callback: (presenceChangedEvent: PresenceEvent) => void
   ): { dispose: () => void };
   /**
-   * Listens to presence changed, the callback will triggered only for passed IDs
-   * @event Listens to presence changed
+   * Listens to presence changes, the callback will triggered only for passed IDs
+   * @event Listens to presence changes
    * @param id contact id (xxxxx@c.us) or group id: xxxxx-yyyy@g.us
    * @param callback Callback of on presence changed
    * @returns Disposable object to stop the listening
@@ -600,7 +626,7 @@ export class ListenerLayer extends ProfileLayer {
   }
 
   /**
-   * @event Listens to revoked messages
+   * @event Listens to message revocation
    * @returns Disposable object to stop the listening
    */
   public onRevokedMessage(
@@ -616,7 +642,7 @@ export class ListenerLayer extends ProfileLayer {
   }
 
   /**
-   * @event Listens to reaction messages
+   * @event Listens to message reactions
    * @returns Disposable object to stop the listening
    */
   public onReactionMessage(